refactor: Address review comments.
