+++ src/entities/join/weighted/validateInputs.ts
