+++ src/entities/types.ts
@@ -1,3 +1,4 @@
+import { MinimalToken } from '../data';
 import { Address, Hex } from '../types';
 
 // Returned from API and used as input
@@ -5,13 +6,7 @@ export type PoolState = {
     id: Hex;
     address: Address;
     type: string;
-    tokens: TokenApi[];
-};
-
-export type TokenApi = {
-    address: Address;
-    decimals: number;
-    index: number;
+    tokens: MinimalToken[];
 };
 
 export type Amounts = {
