+++ src/entities/utils/getSortedTokens.ts
@@ -1,7 +1,10 @@
+import { MinimalToken } from '../../data/types';
 import { Token } from '../token';
-import { TokenApi } from '../types';
 
-export function getSortedTokens(tokens: TokenApi[], chainId: number): Token[] {
+export function getSortedTokens(
+    tokens: MinimalToken[],
+    chainId: number,
+): Token[] {
     return tokens
         .sort((a, b) => a.index - b.index)
         .map((t) => new Token(chainId, t.address, t.decimals));
